+++ tests/Stubs/OCP/AppFramework/Http/Attribute/NoAdminRequired.php
@@ -1,0 +1,6 @@
+<?php
+namespace OCP\AppFramework\Http\Attribute;
+
+#[\Attribute(\Attribute::TARGET_METHOD)]
+class NoAdminRequired {
+}