+++ tests/Stubs/OCP/EventDispatcher/Event.php
@@ -1,0 +1,5 @@
+<?php
+namespace OCP\EventDispatcher;
+
+class Event {
+}