+++ tests/Stubs/OCP/EventDispatcher/IEventDispatcher.php
@@ -1,0 +1,6 @@
+<?php
+namespace OCP\EventDispatcher;
+
+interface IEventDispatcher {
+	public function dispatchTyped(object $event): object;
+}