+++ tests/Stubs/OCP/Files/NotFoundException.php
@@ -1,0 +1,5 @@
+<?php
+namespace OCP\Files;
+
+class NotFoundException extends \Exception {
+}