+++ tests/Stubs/OCP/IAppConfig.php
@@ -1,0 +1,6 @@
+<?php
+namespace OCP;
+
+interface IAppConfig {
+	public function getValueString(string $app, string $key, string $default = ''): string;
+}