+++ tests/Stubs/OCP/IRequest.php
@@ -1,0 +1,5 @@
+<?php
+namespace OCP;
+
+interface IRequest {
+}